some review

+++ app/repository/model/time_all.go
@@ -4,5 +4,5 @@ type TimeAll struct {
 	UUID         string `bson:"uuid"`
 	UserUUID     string `bson:"user_uuid"`
 	CategoryUUID string `bson:"category_uuid"`
-	TimeTotal    int    `bson:"time_total"`
+	TimeTotal    int64  `bson:"time_total"`
 }
